Updated app,js, create_account.html, and created re_create_account.html
Updated app.js to include validation for User ID's when creating an account. If the User ID is not taken, then add the data to the account table and send the postlogin.html file to the client. If the User ID is already taken, then redirect the client to the re_create_account.html file, which works the same as the create_account.html file except that when it loads, the clients gets an alert saying that the User ID is already taken. I left some stuff edited out just in case the changes break the app, so that I can revert them.